add runscripts for main.py and hledger
