+++ run_main.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -euo pipefail
+IFS=$'\n\t'
+
+DEST_DIR="$1" ./docker_run.sh python3 /code/main.py /dest
